Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -54,6 +54,6 @@ def get_average_age_for_season(data, season)
     average_age += age
     number_contestants += 1
   end
-    average_age = average_age.to_i/number_contestants.to_f
+    average_age = average_age.to_i/number_contestants.to_i
     return average_age.ceil
 end
